move vector-related transformations into a separate file
